ATMTests.Tests/ATMTests.cs
using System;
using System.Collections.Generic;
using System.Text;
using ATMCode;
using Xunit;

namespace ATMTests.Tests
{
    public class ATMTests
    {
        [Fact(Skip = "Next test breaks this test.")]
        public static void ViewBalance_should_return_balance()
        {
            // Arrange
            decimal balance = ATM.ViewBalance();

            // Act
            decimal expected = 10.00m;

            // Assert
            Assert.Equal(expected, balance);
        }

        [Fact]
        public static void Withdraw_should_return_requested_amount()
        {
            // Arrange
            decimal withdrawal = 7.50m;

            // Act
            decimal result = ATM.Withdraw(withdrawal);
            decimal currentBalance = ATM.ViewBalance();

            // Assert
            Assert.Equal(7.50m, result);
            Assert.Equal(2.50m, currentBalance);
        }

        [Fact]
        public static void Withdraw_should_return_nothing_if_requested_more_than_balance()
        {
            // Arrange
            decimal withdrawal = 15.00m;

            // Act
            decimal result = ATM.Withdraw(withdrawal);
            decimal currentBalance = ATM.ViewBalance();

            // Assert
            Assert.Equal(0, result);
            Assert.Equal(2.50m, currentBalance);
        }

        [Fact]
        public static void Deposit_should_return_current_balance()
        {
            // Arrange
            decimal depositAmt = 20.00m;

            // Act
            decimal result = ATM.Deposit(depositAmt);

            // Assert
            Assert.Equal(22.50m, result);
        }
    }
}
